Partial clang tidy 14 + refactorings (#106)
* Fix some clang-tidy issues, some warnings
* Add physics frame start and end functions
* remove unneeded `inline` from sort_array.h methods
* use GameEntity based `push_call` in progress_dialog.cpp
* fixed some double/float warnings in editor_spin_slider.cpp
* Remove unused get_closest_points_between_segments
* Use GameEntity based push_call in a few other places
* use range-based for in AudioServer::finish
* reduced nesting in LineEdit::drop_data
* fix warning in generator_helpers.cpp
- remove set_deffered function from Object.h - the function is easily replaced with call_deferred
* Refactor SceneTree group calls + fixes.
* use eastl::sort instead of SortedArray
* code common to call_group_flags, notify_group_flags and set_group_flags moved to call_group_members
* fix comment in physics_2d_server_sw.cpp
* fix some warnings in default_theme.cpp
* fix crash in propagate theme changes control.cpp
* Refactor ProjectSettings
This is mainly to hide pieces of internal logic in the project_settings.cpp from the header.
* small nesting reductions in camera_2d.cpp
* use eastl::sort in cpu_particles_2d.cpp instead of SortedArray + re-
indent.
* Refactor main class cpp file
* Put file-level variables into MainState struct
* extract video/audio driver selection functions
* extract physics processing to `perform_physic_steps`
* extract common cleanup steps to MainState::free_state
* Cleanup in message_queue
- remove `push_set` functions, since they are easily replaced with simple `push_call`
- remove push_call that takes Object * as an argument, now only providing GameEntity based version.
* Remove redundant push_notification from MessageQueue Very easily replaced with `call_deffered([this] {notification(N);} );`

+++ drivers/unix/os_unix.cpp
@@ -517,23 +517,23 @@ void UnixTerminalLogger::log_error(StringView p_function, StringView p_file, int
 
     switch (p_type) {
         case ERR_WARNING:
-            logf_error(FormatVE("%sWARNING:%s %.*s\n", yellow_bold, yellow, err_details.size(),err_details.data()));
+            logf_error(FormatVE("%sWARNING:%s %.*s\n", yellow_bold, yellow, (int)err_details.size(),err_details.data()));
             break;
         case ERR_SCRIPT:
-            logf_error(FormatVE("%sSCRIPT ERROR:%s %.*s\n", magenta_bold, magenta, err_details.size(),err_details.data()));
+            logf_error(FormatVE("%sSCRIPT ERROR:%s %.*s\n", magenta_bold, magenta, (int)err_details.size(),err_details.data()));
             break;
         case ERR_SHADER:
-            logf_error(FormatVE("%sSHADER ERROR:%s %.*s\n", cyan_bold, cyan, err_details.size(),err_details.data()));
+            logf_error(FormatVE("%sSHADER ERROR:%s %.*s\n", cyan_bold, cyan, (int)err_details.size(),err_details.data()));
             break;
         case ERR_ERROR:
         default:
-            logf_error(FormatVE("%sERROR:%s %.*s\n", red_bold, red, err_details.size(),err_details.data()));
+            logf_error(FormatVE("%sERROR:%s %.*s\n", red_bold, red, (int)err_details.size(),err_details.data()));
             break;
     }
     logf_error(FormatVE("%s     at: %.*s (%.*s:%i)%s\n",
                         gray,
-                        p_function.size(),p_function.data(),
-                        p_file.size(), p_file.data(),
+                        (int)p_function.size(),p_function.data(),
+                        (int)p_file.size(), p_file.data(),
                         p_line, reset));
 }
